Support i2c bus other than 1; various spec file improvments
The latest Fedora 39 doesn't load the rtc module the same way anymore.
add an ExecStartPre to handle this as required.

The i2c bus is now 3 in Fedora 39. Add a search for the right bus at
install time. Parameterize the bus number in various places and set the
value in the right files at install time.

Clean up various pre and post directive things to upgrade, install
and remove better.