Handle random enumeration of i2c in Fedora 41 beta
In the latest Fedora 41 beta, the i2c bus enumeration seems to be
random. This causes the discovery at install time to fail randomly. By
migrating the install-time discovery into a script and doing it instead
at system unit start time, it seems to resolve the issue.

There seems to be support for aliases in dtb. Maybe that's a better
approach. We'll look into it. This approach seems robust since the
discovery only uses i2cget, which I think is safe.